// frontend/src/app/core/generated/circabc/model/searchConfig.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { KeywordDefinition } from './keywordDefinition';


export interface SearchConfig { 
    searchName?: string;
    searchFor?: string;
    searchIn?: SearchConfig.SearchInEnum;
    language?: string;
    creationDateFrom?: string;
    creationDateTo?: string;
    lastModificationFrom?: string;
    lastModificationTo?: string;
    creatorUser?: User;
    keywords?: Array<KeywordDefinition>;
    status?: SearchConfig.StatusEnum;
}
export namespace SearchConfig {
    export const SearchInEnum = {
        All: 'All',
        Library: 'library',
        Forums: 'forums',
        InfoService: 'infoService',
        Agenda: 'agenda'
    } as const;
    export type SearchInEnum = typeof SearchInEnum[keyof typeof SearchInEnum];
    export const StatusEnum = {
        All: 'All',
        Library: 'library',
        Forums: 'forums',
        InfoService: 'infoService',
        Agenda: 'agenda'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}