// frontend/src/app/core/generated/circabc/model/searchPreferences.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SearchConfig } from './searchConfig';


export interface SearchPreferences { 
    igId: string;
    searchConfig: Array<SearchConfig>;
}